refactor: split RecognizeCallback into abstract base and concrete enroll/authenticate callbacks
