+++ Recognize/Recognize/Journey/RecognizeCallback.swift
@@ -10,528 +10,41 @@
 
 #if canImport(KeylessSDK)
 import Foundation
-@preconcurrency import KeylessSDK
-import PingOrchestrate
 import PingJourneyPlugin
 
-
-// MARK: - RecognizeOperationType
-
-/// The operation type conveyed by the server in the `operationType` output field.
-public enum RecognizeOperationType: String, Sendable {
-    case enroll = "ENROLL"
-    case authenticate = "AUTHENTICATE"
-}
-
-// MARK: - RecognizeMobileSDKOptions
-
-/// Typed wrapper around the `mobileSDKOptions` dictionary received from the server.
+/// Factory callback registered for the `"PingOneRecognizeCallback"` server type.
 ///
-/// The server delivers all values as strings, including booleans (`"true"`/`"false"`) and
-/// integers (`"2"`, `"5"`). Computed properties convert these to native Swift types.
-public struct RecognizeMobileSDKOptions: Sendable {
-
-    let raw: [String: String]
-
-    // MARK: Common options
-
-    /// The liveness detection configuration level (e.g. `"LEVEL_1"`).
-    public var livenessConfiguration: String { raw[JourneyConstants.livenessConfiguration] ?? "" }
-
-    /// Whether the SDK adapts its behaviour to the surrounding environment.
-    public var livenessEnvironmentAware: Bool? { raw[JourneyConstants.livenessEnvironmentAware].flatMap(Bool.init) }
-
-    /// The server-assigned operation info ID.
-    public var operationInfoId: String { raw[JourneyConstants.operationInfoId] ?? "" }
-
-    /// The server-assigned operation info payload.
-    public var operationInfoPayload: String { raw[JourneyConstants.operationInfoPayload] ?? "" }
-
-    /// The external user ID linked to this operation.
-    public var operationInfoExternalUserId: String { raw[JourneyConstants.operationInfoExternalUserId] ?? "" }
-
-    /// Delay in seconds before the camera activates.
-    public var cameraDelaySeconds: Int? { raw[JourneyConstants.cameraDelaySeconds].flatMap(Int.init) }
-
-    /// Whether to show a success feedback overlay after the operation completes.
-    public var showSuccessFeedback: Bool? { raw[JourneyConstants.showSuccessFeedback].flatMap(Bool.init) }
-
-    // MARK: Enrollment-only options
-
-    /// A custom secret to bind to the enrollment record.
-    /// - Note: Parsed from the server but not yet forwarded to the Keyless SDK — behaviour under investigation.
-    public var customSecret: String { raw[JourneyConstants.customSecret] ?? "" }
-
-    /// Whether to show a failure feedback overlay when enrollment fails.
-    public var showFailureFeedback: Bool? { raw[JourneyConstants.showFailureFeedback].flatMap(Bool.init) }
-
-    /// Whether to display the instructions screen before the camera activates.
-    public var showInstructionsScreen: Bool? { raw[JourneyConstants.showInstructionsScreen].flatMap(Bool.init) }
-
-    /// The UI presentation style for enrollment (e.g. `"OVERLAY"`).
-    public var presentation: String { raw[JourneyConstants.presentation] ?? "" }
-
-    /// The number of enrollment circuits the SDK runs to capture biometrics.
-    public var numberOfEnrollmentCircuits: Int? { raw[JourneyConstants.numberOfEnrollmentCircuits].flatMap(Int.init) }
-
-    // MARK: Authentication-only options
-
-    /// The UI presentation style for authentication (e.g. `"CAMERA_PREVIEW"`).
-    public var presentationStyle: String { raw[JourneyConstants.presentationStyle] ?? "" }
-
-}
-
-// MARK: - RecognizeCallback
-
-/// A Journey callback that drives the PingOne Recognize biometric authentication SDK.
+/// The server emits a single callback type for both enrollment and authentication.
+/// `RecognizeCallback` reads the `operationType` output field during `initialize(with:)`
+/// and returns the appropriate concrete instance:
+/// - `ENROLL` → `PingOneRecognizeEnrollCallback`
+/// - `AUTHENTICATE` (with clientState) → `PingOneRecognizeEnrollCallback` (enrollment-restore path)
+/// - `AUTHENTICATE` (without clientState) → `PingOneRecognizeAuthenticateCallback`
 ///
-/// The server emits a single callback type (`"PingOneRecognizeCallback"`) for both
-/// enrollment and authentication. The `operationType` output field distinguishes the two flows.
-///
-/// **Usage:**
-/// ```swift
-/// if let recognizeCallback = node.callbacks.first(where: { $0 is RecognizeCallback })
-///         as? RecognizeCallback {
-///     let result = await recognizeCallback.execute()
-/// }
-/// ```
-public class RecognizeCallback: AbstractCallback, ContinueNodeAware, @unchecked Sendable {
+/// This class is not intended to be used directly — cast the result of
+/// `node.callbacks` to `PingOneRecognizeEnrollCallback` or `PingOneRecognizeAuthenticateCallback`.
+class RecognizeCallback: AbstractRecognizeCallback, @unchecked Sendable {
 
-    /// Reference to the continue node for accessing other callbacks.
-    public weak var continueNode: ContinueNode?
+    override func initialize(with json: [String: Any]) async -> any Callback {
+        // Parse shared output fields into self first.
+        _ = await super.initialize(with: json)
 
-    // MARK: - Output Properties
-
-    /// Whether this is an enrollment or authentication operation.
-    private(set) public var operationType: RecognizeOperationType = .enroll
-
-    /// The WebSocket URL for the Recognize authentication service.
-    /// - Note: Provided by the server for context but not forwarded to the Keyless SDK, which manages its own transport layer.
-    private(set) public var websocketURL: String = ""
-
-    /// The customer name associated with this PingOne Recognize deployment.
-    /// - Note: Provided by the server for context but not forwarded to the Keyless SDK.
-    private(set) public var customerName: String = ""
-
-    /// The public key used to encrypt images before sending them to the server.
-    /// - Note: Provided by the server for context but not forwarded to the Keyless SDK, which handles image encryption internally.
-    private(set) public var imageEncryptionPublicKey: String = ""
-
-    /// The ID that identifies the image encryption key.
-    /// - Note: Provided by the server for context but not forwarded to the Keyless SDK.
-    private(set) public var imageEncryptionKeyId: String = ""
-
-    /// The Recognize node host URL (used to configure the Keyless SDK).
-    private(set) public var host: String = ""
-
-    /// The API key for the Keyless SDK.
-    private(set) public var apiKey: String = ""
-
-    /// The username associated with the biometric operation.
-    /// - Note: Provided by the server for context but not forwarded to the Keyless SDK.
-    ///   User identity is bound via `operationInfoExternalUserId` when the server supplies an explicit `operationInfoId`.
-    private(set) public var username: String = ""
-
-    /// Customer-supplied transaction data to be signed by the SDK.
-    private(set) public var transactionData: String = ""
-
-    /// Whether the server requested generation of a new client state.
-    private(set) public var generateClientState: Bool = false
-
-    /// An existing client-state payload provided by the server during enrollment restore.
-    private(set) public var clientState: String = ""
-
-    /// Mobile SDK options parsed from the server response.
-    private(set) public var mobileSDKOptions: RecognizeMobileSDKOptions = RecognizeMobileSDKOptions(raw: [:])
-
-    // MARK: - Initialization
-
-    /// Populates output properties from the JSON output fields delivered by the server.
-    public override func initValue(name: String, value: Any) {
-        switch name {
-        case JourneyConstants.operationType:
-            if let stringValue = value as? String {
-                self.operationType = RecognizeOperationType(rawValue: stringValue) ?? .enroll
-            }
-        case JourneyConstants.websocketURL:
-            if let stringValue = value as? String { self.websocketURL = stringValue }
-        case JourneyConstants.customerName:
-            if let stringValue = value as? String { self.customerName = stringValue }
-        case JourneyConstants.imageEncryptionPublicKey:
-            if let stringValue = value as? String { self.imageEncryptionPublicKey = stringValue }
-        case JourneyConstants.imageEncryptionKeyId:
-            if let stringValue = value as? String { self.imageEncryptionKeyId = stringValue }
-        case JourneyConstants.host:
-            if let stringValue = value as? String { self.host = stringValue }
-        case JourneyConstants.apiKey:
-            if let stringValue = value as? String { self.apiKey = stringValue }
-        case JourneyConstants.username:
-            if let stringValue = value as? String { self.username = stringValue }
-        case JourneyConstants.transactionData:
-            if let stringValue = value as? String { self.transactionData = stringValue }
-        case JourneyConstants.generateClientState:
-            if let boolValue = value as? Bool {
-                self.generateClientState = boolValue
-            } else if let stringValue = value as? String {
-                self.generateClientState = Bool(stringValue) ?? false
-            }
-        case JourneyConstants.clientState:
-            if let stringValue = value as? String { self.clientState = stringValue }
-        case JourneyConstants.mobileSDKOptions:
-            self.mobileSDKOptions = RecognizeMobileSDKOptions(raw: Self.stringDictionary(from: value))
-        default:
-            break
-        }
-    }
-
-    // MARK: - Execution
-
-    /// Configures the Keyless SDK and executes the enrollment or authentication operation.
-    ///
-    /// - Returns: `.success(())` on completion, or `.failure(error)` if any step fails.
-    ///   On failure the `IDToken1clientError` input field is automatically populated.
-    public func execute() async -> Result<Void, Error> {
-        do {
-            try await configure()
-            try Task.checkCancellation()
-            switch operationType {
-            case .enroll:
-                try await enroll(options: mobileSDKOptions)
-            case .authenticate:
-                if !clientState.isEmpty {
-                    try await enrollWithClientState(clientState, options: mobileSDKOptions)
-                } else {
-                    try await authenticate(options: mobileSDKOptions)
-                }
-            }
-            return .success(())
-        } catch {
-            let message = error.localizedDescription.isEmpty
-                ? JourneyConstants.clientError
-                : error.localizedDescription
-            self.error(message)
-            if let recognizeError = error as? RecognizeError {
-                setClientErrorCode(String(recognizeError.code))
-            }
-            return .failure(error)
-        }
-    }
-
-    // MARK: - Input Helpers
-
-    /// Writes the error message into the `IDToken1clientError` input field.
-    /// - Parameter value: The error message to send to the server.
-    public func error(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputClientError)
-    }
-
-    /// Sets the signed JWT returned by the Keyless SDK into the `IDToken1signedJwt` input field.
-    public func setSignedJwt(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputSignedJwt)
-    }
-
-    /// Sets the client state returned by the Keyless SDK into the `IDToken1clientState` input field.
-    public func setClientState(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputClientState)
-    }
-
-    /// Sets the Recognize user ID into the `IDToken1recognizeId` input field.
-    public func setRecognizeId(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputRecognizeId)
-    }
-
-    /// Sets the device public signing key into the `IDToken1devicePublicSigningKey` input field.
-    public func setDevicePublicSigningKey(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputDevicePublicSigningKey)
-    }
-
-    /// Sets a client error code into the `IDToken1clientErrorCode` input field.
-    public func setClientErrorCode(_ value: String) {
-        _ = input(value, forKey: JourneyConstants.inputClientErrorCode)
-    }
-
-    // MARK: - Keyless SDK Wrappers
-
-    /// Returns a `JwtSigningInfo` when `transactionData` is non-empty; `nil` otherwise.
-    private func jwtSigningInfo(from transactionData: String) -> JwtSigningInfo? {
-        transactionData.isEmpty ? nil : JwtSigningInfo(claimTransactionData: transactionData)
-    }
-
-    /// Configures and initialises the Keyless SDK with the server-supplied credentials.
-    internal func configure() async throws {
-        let setupConfig = SetupConfig(
-            apiKey: apiKey,
-            hosts: [host],
-            numberOfEnrollmentCircuits: mobileSDKOptions.numberOfEnrollmentCircuits ?? SetupConfig.DEFAULT_NUMBER_OF_ENROLLMENT_CIRCUITS
-        )
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            Keyless.configure(setupConfiguration: setupConfig) { error in
-                if let error = error {
-                    continuation.resume(throwing: RecognizeError(error.message, code: error.code))
-                } else {
-                    continuation.resume()
-                }
+        switch operationType {
+        case .enroll:
+            let callback = PingOneRecognizeEnrollCallback()
+            _ = await callback.initialize(with: json)
+            return callback
+        case .authenticate:
+            if !clientState.isEmpty {
+                let callback = PingOneRecognizeEnrollCallback()
+                _ = await callback.initialize(with: json)
+                return callback
+            } else {
+                let callback = PingOneRecognizeAuthenticateCallback()
+                _ = await callback.initialize(with: json)
+                return callback
             }
         }
     }
-
-    /// Performs the biometric enrollment operation using `BiomEnrollConfig`.
-    ///
-    /// On success, populates `IDToken1recognizeId`, `IDToken1signedJwt`, and `IDToken1clientState`.
-    internal func enroll(options: RecognizeMobileSDKOptions) async throws {
-        let operationInfo: Keyless.OperationInfo? = options.operationInfoId.isEmpty ? nil
-            : Keyless.OperationInfo(
-                id: options.operationInfoId,
-                payload: options.operationInfoPayload,
-                externalUserId: options.operationInfoExternalUserId.isEmpty ? nil : options.operationInfoExternalUserId
-            )
-
-        let enrollConfig = BiomEnrollConfig(
-            clientState: clientState.isEmpty ? nil : clientState,
-            operationInfo: operationInfo,
-            jwtSigningInfo: jwtSigningInfo(from: transactionData),
-            livenessConfiguration: Self.livenessConfiguration(from: options.livenessConfiguration),
-            // BiomEnrollConfig has no DEFAULT_LIVENESS_ENV_AWARE; BiomAuthConfig's constant carries the same value.
-            livenessEnvironmentAware: options.livenessEnvironmentAware ?? BiomAuthConfig.DEFAULT_LIVENESS_ENV_AWARE,
-            cameraDelaySeconds: options.cameraDelaySeconds ?? BiomEnrollConfig.DEFAULT_DELAY,
-            generatingClientState: generateClientState ? .backup : nil,
-            showInstructionsScreen: options.showInstructionsScreen ?? BiomEnrollConfig.DEFAULT_SHOW_INSTRUCTIONS_SCREEN,
-            showSuccessFeedback: options.showSuccessFeedback ?? BiomEnrollConfig.DEFAULT_SHOW_SUCCESS_FEEDBACK,
-            showFailureFeedback: options.showFailureFeedback ?? BiomEnrollConfig.DEFAULT_SHOW_FAILURE_FEEDBACK,
-            presentationStyle: Self.enrollPresentationStyle(from: options.presentation)
-        )
-
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            Keyless.enroll(configuration: enrollConfig) { [weak self] result in
-                switch result {
-                case .success(let enrollmentResult):
-                    if let keylessId = enrollmentResult.keylessId { self?.setRecognizeId(keylessId) }
-                    if let jwt = enrollmentResult.signedJwt { self?.setSignedJwt(jwt) }
-                    if let state = enrollmentResult.clientState { self?.setClientState(state) }
-                    if case .success(let key) = Keyless.getDevicePublicSigningKey() {
-                        self?.setDevicePublicSigningKey(key)
-                    }
-                    continuation.resume()
-                case .failure(let error):
-                    if let sdkError = error as? KeylessSDKError {
-                        continuation.resume(throwing: RecognizeError(sdkError.message, code: sdkError.code))
-                    } else {
-                        continuation.resume(throwing: error)
-                    }
-                }
-            }
-        }
-    }
-
-    /// Performs the biometric authentication operation using `BiomAuthConfig`.
-    ///
-    /// On success, populates `IDToken1signedJwt`, `IDToken1clientState`, and (when available)
-    /// `IDToken1devicePublicSigningKey`.
-    internal func authenticate(options: RecognizeMobileSDKOptions) async throws {
-        let operationInfo: Keyless.OperationInfo? = options.operationInfoId.isEmpty ? nil
-            : Keyless.OperationInfo(
-                id: options.operationInfoId,
-                payload: options.operationInfoPayload,
-                externalUserId: options.operationInfoExternalUserId.isEmpty ? nil : options.operationInfoExternalUserId
-            )
-
-        let authConfig = BiomAuthConfig(
-            livenessConfiguration: Self.livenessConfiguration(from: options.livenessConfiguration),
-            livenessEnvironmentAware: options.livenessEnvironmentAware ?? BiomAuthConfig.DEFAULT_LIVENESS_ENV_AWARE,
-            cameraDelaySeconds: options.cameraDelaySeconds ?? BiomAuthConfig.DEFAULT_CAMERA_DELAY_SECONDS,
-            generatingClientState: generateClientState ? .backup : nil,
-            showSuccessFeedback: options.showSuccessFeedback ?? BiomAuthConfig.DEFAULT_SHOW_SUCCESS_FEEDBACK,
-            jwtSigningInfo: jwtSigningInfo(from: transactionData),
-            operationInfo: operationInfo,
-            presentationStyle: Self.authPresentationStyle(from: options.presentationStyle)
-        )
-
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            Keyless.authenticate(configuration: authConfig) { [weak self] result in
-                switch result {
-                case .success(let response):
-                    if let jwt = response.signedJwt { self?.setSignedJwt(jwt) }
-                    if let state = response.clientState { self?.setClientState(state) }
-                    if case .success(let key) = Keyless.getDevicePublicSigningKey() {
-                        self?.setDevicePublicSigningKey(key)
-                    }
-                    continuation.resume()
-                case .failure(let error):
-                    if let sdkError = error as? KeylessSDKError {
-                        continuation.resume(throwing: RecognizeError(sdkError.message, code: sdkError.code))
-                    } else {
-                        continuation.resume(throwing: error)
-                    }
-                }
-            }
-        }
-    }
-
-    /// Handles the authenticate-with-clientState path: checks whether the user is already enrolled.
-    ///
-    /// Called from `execute()` in place of `authenticate()` when the server provides a `clientState`.
-    /// - If `validateUserDeviceActive` returns nil the user is already enrolled — runs normal authentication.
-    /// - If `validateUserDeviceActive` returns `userNotEnrolled` — calls `enroll(clientState:)`.
-    /// - Any other error from either call is propagated as a `RecognizeError`.
-    internal func enrollWithClientState(_ clientState: String, options: RecognizeMobileSDKOptions) async throws {
-        // Step 1: validate — nil means already enrolled (skip), userNotEnrolled means proceed, anything else is an error.
-        let validationError: KeylessSDKError? = await withCheckedContinuation { continuation in
-            Keyless.validateUserDeviceActive { error in
-                continuation.resume(returning: error)
-            }
-        }
-
-        if let error = validationError {
-            guard case .integrationError(let ie) = error.kind, ie == .userNotEnrolled else {
-                throw RecognizeError(error.message, code: error.code)
-            }
-            // userNotEnrolled — fall through to enroll below.
-        } else {
-            // nil — user is already enrolled, run normal authentication.
-            try await authenticate(options: options)
-            return
-        }
-
-        // Step 2: user is not enrolled — enroll with the server-supplied clientState.
-        let enrollConfig = BiomEnrollConfig(
-            clientState: clientState,
-            livenessConfiguration: Self.livenessConfiguration(from: options.livenessConfiguration),
-            // BiomEnrollConfig has no DEFAULT_LIVENESS_ENV_AWARE; BiomAuthConfig's constant carries the same value.
-            livenessEnvironmentAware: options.livenessEnvironmentAware ?? BiomAuthConfig.DEFAULT_LIVENESS_ENV_AWARE,
-            cameraDelaySeconds: options.cameraDelaySeconds ?? BiomEnrollConfig.DEFAULT_DELAY,
-            generatingClientState: generateClientState ? .backup : nil,
-            showInstructionsScreen: options.showInstructionsScreen ?? BiomEnrollConfig.DEFAULT_SHOW_INSTRUCTIONS_SCREEN,
-            showSuccessFeedback: options.showSuccessFeedback ?? BiomEnrollConfig.DEFAULT_SHOW_SUCCESS_FEEDBACK,
-            showFailureFeedback: options.showFailureFeedback ?? BiomEnrollConfig.DEFAULT_SHOW_FAILURE_FEEDBACK,
-            presentationStyle: Self.enrollPresentationStyle(from: options.presentation)
-        )
-
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            Keyless.enroll(configuration: enrollConfig) { [weak self] result in
-                switch result {
-                case .success(let enrollmentResult):
-                    if let keylessId = enrollmentResult.keylessId { self?.setRecognizeId(keylessId) }
-                    if let jwt = enrollmentResult.signedJwt { self?.setSignedJwt(jwt) }
-                    if let state = enrollmentResult.clientState { self?.setClientState(state) }
-                    if case .success(let key) = Keyless.getDevicePublicSigningKey() {
-                        self?.setDevicePublicSigningKey(key)
-                    }
-                    continuation.resume()
-                case .failure(let error):
-                    continuation.resume(throwing: RecognizeError(error.message, code: error.code))
-                }
-            }
-        }
-    }
-
-    // MARK: - Private Helpers
-
-    /// Maps the server liveness string (e.g. `"LEVEL_1"`) to `Keyless.LivenessConfiguration`.
-    ///
-    /// Falls back to `.LEVEL_1` for unrecognised values.
-    private static func livenessConfiguration(from string: String) -> Keyless.LivenessConfiguration {
-        switch string.uppercased() {
-        case "DEVELOPMENT": return .DEVELOPMENT
-        case "LEVEL_1":     return .LEVEL_1
-        case "LEVEL_2":     return .LEVEL_2
-        default:            return .LEVEL_1
-        }
-    }
-
-    /// Maps the server `presentation` string (e.g. `"OVERLAY"`) to `BiomEnrollConfig.PresentationStyle`.
-    ///
-    /// Falls back to `.overlay` for unrecognised values.
-    private static func enrollPresentationStyle(from string: String) -> BiomEnrollConfig.PresentationStyle {
-        switch string.uppercased() {
-        case "FULL_SCREEN": return .fullScreen
-        default:            return .overlay
-        }
-    }
-
-    /// Maps the server `presentationStyle` string (e.g. `"CAMERA_PREVIEW"`) to `PresentationStyle`.
-    ///
-    /// Falls back to `.cameraPreview` for unrecognised values.
-    private static func authPresentationStyle(from string: String) -> PresentationStyle {
-        switch string.uppercased() {
-        case "NO_CAMERA_PREVIEW": return .noCameraPreview
-        default:                  return .cameraPreview
-        }
-    }
-
-    /// Converts a JSON value to a `[String: String]` dictionary.
-    ///
-    /// Values that are not already strings are coerced via string interpolation.
-    private static func stringDictionary(from value: Any) -> [String: String] {
-        if let dict = value as? [String: String] { return dict }
-        if let dict = value as? [String: Any] {
-            return dict.reduce(into: [:]) { result, pair in
-                result[pair.key] = "\(pair.value)"
-            }
-        }
-        return [:]
-    }
-}
-
-// MARK: - JourneyConstants (Recognize input field keys)
-
-extension JourneyConstants {
-    /// Input field key for the signed JWT produced by the Keyless SDK.
-    public static let inputSignedJwt = "IDToken1signedJwt"
-    /// Input field key for the client state produced by the Keyless SDK.
-    public static let inputClientState = "IDToken1clientState"
-    /// Input field key for the Recognize user ID produced during enrollment.
-    public static let inputRecognizeId = "IDToken1recognizeId"
-    /// Input field key for the device public signing key retrieved after authentication.
-    public static let inputDevicePublicSigningKey = "IDToken1devicePublicSigningKey"
-    /// Input field key for the client error message.
-    public static let inputClientError = "IDToken1clientError"
-    /// Input field key for a structured client error code.
-    public static let inputClientErrorCode = "IDToken1clientErrorCode"
-}
-
-// MARK: - JourneyConstants (Recognize output field keys + defaults)
-
-extension JourneyConstants {
-    /// Fallback error string used when the error localised description is empty.
-    public static let clientError = "clientError"
-
-    // MARK: PingOneRecognize — mobile SDK option default values
-
-    /// String representation of a `true` boolean as delivered by the server.
-    public static let boolTrue = "true"
-    /// String representation of a `false` boolean as delivered by the server.
-    public static let boolFalse = "false"
-
-    // MARK: PingOneRecognize — output field keys
-
-    public static let operationType = "operationType"
-    public static let websocketURL = "websocketURL"
-    public static let customerName = "customerName"
-    public static let imageEncryptionPublicKey = "imageEncryptionPublicKey"
-    public static let imageEncryptionKeyId = "imageEncryptionKeyId"
-    public static let host = "host"
-    public static let apiKey = "apiKey"
-    public static let username = "username"
-    public static let transactionData = "transactionData"
-    public static let generateClientState = "generateClientState"
-    public static let clientState = "clientState"
-    public static let mobileSDKOptions = "mobileSDKOptions"
-    public static let webSDKOptions = "webSDKOptions"
-
-    // MARK: PingOneRecognize — mobile SDK option keys
-
-    public static let livenessConfiguration = "livenessConfiguration"
-    public static let livenessEnvironmentAware = "livenessEnvironmentAware"
-    public static let operationInfoId = "operationInfoId"
-    public static let operationInfoPayload = "operationInfoPayload"
-    public static let operationInfoExternalUserId = "operationInfoExternalUserId"
-    public static let cameraDelaySeconds = "cameraDelaySeconds"
-    public static let showSuccessFeedback = "showSuccessFeedback"
-    public static let customSecret = "customSecret"
-    public static let showFailureFeedback = "showFailureFeedback"
-    public static let showInstructionsScreen = "showInstructionsScreen"
-    public static let presentation = "presentation"
-    public static let numberOfEnrollmentCircuits = "numberOfEnrollmentCircuits"
-    public static let presentationStyle = "presentationStyle"
 }
 #endif
